requirements.txt
# python -m pip install --upgrade -r requirements.txt
# python -m pip list --outdated
faiss-cpu==1.11.0
# or faiss-gpu
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.23
openai==1.86.0
python-dotenv==1.1.0
tiktoken==0.9.0
youtube-transcript-api==1.1.0
# Development/Build
pex==2.1.141
